requirements.txt
transformers
datasets
mlflow
torch
pandas
scikit-learn
sqlalchemy
gunicorn
gradio
pyngrok 
pyyaml
dvc[s3]